requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.9
django-allauth==0.60.1
django-cors-headers==4.3.1
django-import-export==3.3.6
et-xmlfile==1.1.0
idna==3.6
MarkupPy==1.14
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
